+++ collector/src/main/java/com/bitmovin/analytics/BitmovinAnalyticsConfig.java
@@ -3,6 +3,7 @@ package com.bitmovin.analytics;
 import android.content.Context;
 import android.os.Parcel;
 import android.os.Parcelable;
+import com.bitmovin.analytics.data.CustomData;
 import com.bitmovin.analytics.enums.CDNProvider;
 import com.bitmovin.analytics.enums.PlayerType;
 
@@ -31,18 +32,17 @@ public class BitmovinAnalyticsConfig implements Parcelable {
     private Boolean isLive;
     private CollectorConfig config = new CollectorConfig();
 
-    public static final Creator<BitmovinAnalyticsConfig> CREATOR =
-            new Creator<BitmovinAnalyticsConfig>() {
-                @Override
-                public BitmovinAnalyticsConfig createFromParcel(Parcel in) {
-                    return new BitmovinAnalyticsConfig(in);
-                }
+    public static final Creator<BitmovinAnalyticsConfig> CREATOR = new Creator<BitmovinAnalyticsConfig>() {
+        @Override
+        public BitmovinAnalyticsConfig createFromParcel(Parcel in) {
+            return new BitmovinAnalyticsConfig(in);
+        }
 
-                @Override
-                public BitmovinAnalyticsConfig[] newArray(int size) {
-                    return new BitmovinAnalyticsConfig[size];
-                }
-            };
+        @Override
+        public BitmovinAnalyticsConfig[] newArray(int size) {
+            return new BitmovinAnalyticsConfig[size];
+        }
+    };
 
     @Deprecated
     public BitmovinAnalyticsConfig(String key, Context context) {
@@ -91,7 +91,8 @@ public class BitmovinAnalyticsConfig implements Parcelable {
         ads = in.readInt() == 1;
     }
 
-    public BitmovinAnalyticsConfig() {}
+    public BitmovinAnalyticsConfig() {
+    }
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
@@ -241,7 +242,8 @@ public class BitmovinAnalyticsConfig implements Parcelable {
     }
 
     /**
-     * Optional free-form data Not enabled by default Must be activated for your organization
+     * Optional free-form data Not enabled by default Must be activated for your
+     * organization
      *
      * @param customData6
      */
@@ -254,7 +256,8 @@ public class BitmovinAnalyticsConfig implements Parcelable {
     }
 
     /**
-     * Optional free-form data Not enabled by default Must be activated for your organization
+     * Optional free-form data Not enabled by default Must be activated for your
+     * organization
      *
      * @param customData7
      */
@@ -290,8 +293,8 @@ public class BitmovinAnalyticsConfig implements Parcelable {
     }
 
     /**
-     * Set MPD URL recorded in analytics. If not set explicitly the collector will retrieve
-     * available information from the player.
+     * Set MPD URL recorded in analytics. If not set explicitly the collector will
+     * retrieve available information from the player.
      *
      * @param mpdUrl
      */
@@ -304,8 +307,8 @@ public class BitmovinAnalyticsConfig implements Parcelable {
     }
 
     /**
-     * Set M3U8 URL recorded in analytics. If not set explicitly the collector will retrieve
-     * available information from the player.
+     * Set M3U8 URL recorded in analytics. If not set explicitly the collector will
+     * retrieve available information from the player.
      *
      * @param m3u8Url
      */
@@ -388,7 +391,8 @@ public class BitmovinAnalyticsConfig implements Parcelable {
     }
 
     /**
-     * Returns true if the stream is marked as live before stream metadata is available.
+     * Returns true if the stream is marked as live before stream metadata is
+     * available.
      *
      * @return
      */
@@ -403,5 +407,22 @@ public class BitmovinAnalyticsConfig implements Parcelable {
      */
     public void setIsLive(Boolean live) {
         isLive = live;
+    }
+
+    protected CustomData getCustomData() {
+        return new CustomData(this.getCustomData1(), this.getCustomData2(), this.getCustomData3(),
+                this.getCustomData4(), this.getCustomData5(), this.getCustomData6(), this.getCustomData7(),
+                this.getExperimentName());
+    }
+
+    protected void setCustomData(CustomData customData) {
+        this.setCustomData1(customData.getCustomData1());
+        this.setCustomData2(customData.getCustomData2());
+        this.setCustomData3(customData.getCustomData3());
+        this.setCustomData4(customData.getCustomData4());
+        this.setCustomData5(customData.getCustomData5());
+        this.setCustomData6(customData.getCustomData6());
+        this.setCustomData7(customData.getCustomData7());
+        this.setExperimentName(customData.getExperimentName());
     }
 }