+++ collector/src/main/java/com/bitmovin/analytics/data/CustomData.kt
@@ -1,0 +1,12 @@
+package com.bitmovin.analytics.data
+
+data class CustomData(
+    var customData1: String? = null,
+    var customData2: String? = null,
+    var customData3: String? = null,
+    var customData4: String? = null,
+    var customData5: String? = null,
+    var customData6: String? = null,
+    var customData7: String? = null,
+    var experimentName: String? = null
+)